Move function `verify_url_from_reference`
