+++ warehouse/oidc/models/github.py
@@ -33,8 +33,8 @@ from warehouse.oidc.models._core import (
     PendingOIDCPublisher,
     check_claim_binary,
     check_existing_jti,
-    verify_url_from_reference,
 )
+from warehouse.oidc.urls import verify_url_from_reference
 
 GITHUB_OIDC_ISSUER_URL = "https://token.actions.githubusercontent.com"
